+++ apps/data-service/src/services/pairs/repo/transformResult.ts
@@ -13,7 +13,7 @@ export type PairDbResponse = {
   low: BigNumber;
   weighted_average_price: BigNumber;
   txs_count: number;
-  volume_waves: BigNumber;
+  volume_dcc: BigNumber;
 };
 
 /** renamePairFields :: Object -> Object */
@@ -24,7 +24,7 @@ const renamePairFields = renameKeys<PairInfo & AssetIdsPair, PairDbResponse>({
   price_asset_id: 'priceAsset',
   quote_volume: 'quoteVolume',
   txs_count: 'txsCount',
-  volume_waves: 'volumeWaves',
+  volume_dcc: 'volumeDcc',
   weighted_average_price: 'weightedAveragePrice',
 });