+++ packages/ride/lang/shared/src/main/scala/com/decentralchain/lang/v1/evaluator/ctx/impl/waves/package.scala
@@ -1,6 +1,0 @@
-package com.decentralchain.lang.v1.evaluator.ctx.impl
-
-package object waves {
-  val ExtractedFuncPrefix  = "@extr"
-  val ExtractedFuncPostfix = "Value"
-}